backend/requirements.txt

fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
openai==1.52.0
python-dotenv==1.0.1
supabase==2.25.0
httpx==0.27.2
pydantic==2.11.7
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
loguru==0.7.2
reportlab>=4.0.0